ui/properties: extract search bar tab with section helpers
