+++ panel-plugin/meson.build
@@ -37,6 +37,9 @@ plugin_sources = [
   'ui/properties/common.cpp',
   'ui/properties/general.cpp',
   'ui/properties/user-session.cpp',
+  'ui/properties/search-bar.cpp',
+  'ui/properties/search-bar-aliases.cpp',
+  'ui/properties/search-bar-actions.cpp',
   'slot.h',
   'window.cpp',
   xfce_revision_h,
