+++ panel-plugin/settings-dialog.cpp
@@ -732,523 +732,11 @@ void SettingsDialog::apply_unified_bar_sensitivity()
 }
 
 //-----------------------------------------------------------------------------
-// New 5-tab Properties dialog — empty stubs, filled by Phase 4–8 user stories.
+// Per-tab builders for the Properties dialog. Each init_*_tab() returns a
+// fully wired scrolled container ready to be added to the dialog's stack.
+// Builders for the General, User/Session and Search Bar tabs live in their
+// own translation units under ui/properties/.
 //-----------------------------------------------------------------------------
-
-/* init_search_bar_tab:
- *
- * Builds the Search Bar tab in the 003-properties-refactor 5-tab dictionary.
- * Sections (top-to-bottom, FR-030):
- *   1. Position        — search-bar-position combo.
- *   2. Ranking         — fuzzy matching, favorites boost, recency weight
- *                        (lifted from the legacy "Advanced Search" tab).
- *   3. Aliases         — desktop-id → search-term map.
- *   4. Search actions  — list of user-defined actions with add/remove/edit;
- *                        the Edit button opens a transient modal per
- *                        research R-7 / data-model E-5.
- *
- * Returns: a scrolled container ready to be packed into the dialog's stack.
- */
-GtkWidget* SettingsDialog::init_search_bar_tab()
-{
-	GtkBox* page = GTK_BOX(gtk_box_new(GTK_ORIENTATION_VERTICAL, 18));
-	gtk_container_set_border_width(GTK_CONTAINER(page), 12);
-
-	GtkSizeGroup* label_size_group = gtk_size_group_new(GTK_SIZE_GROUP_HORIZONTAL);
-	GtkSizeGroup* control_size_group = gtk_size_group_new(GTK_SIZE_GROUP_HORIZONTAL);
-
-	// =========================================================================
-	// 1. Position section
-	// =========================================================================
-	GtkGrid* pos_table = GTK_GRID(gtk_grid_new());
-	gtk_grid_set_column_spacing(pos_table, 12);
-	gtk_grid_set_row_spacing(pos_table, 6);
-
-	GtkWidget* pos_frame = make_aligned_frame(_("Position"), GTK_WIDGET(pos_table));
-	gtk_box_pack_start(page, pos_frame, false, false, 0);
-
-	GtkWidget* sbp_label = gtk_label_new_with_mnemonic(_("_Search bar position:"));
-	gtk_widget_set_halign(sbp_label, GTK_ALIGN_START);
-	gtk_grid_attach(pos_table, sbp_label, 0, 0, 1, 1);
-
-	m_search_bar_position_combo = gtk_combo_box_text_new();
-	gtk_combo_box_text_append(GTK_COMBO_BOX_TEXT(m_search_bar_position_combo), "top", _("Top"));
-	gtk_combo_box_text_append(GTK_COMBO_BOX_TEXT(m_search_bar_position_combo), "bottom", _("Bottom"));
-	gtk_combo_box_set_active_id(GTK_COMBO_BOX(m_search_bar_position_combo),
-		static_cast<const gchar*>(m_settings->search_bar_position));
-	gtk_widget_set_halign(m_search_bar_position_combo, GTK_ALIGN_START);
-	gtk_grid_attach(pos_table, m_search_bar_position_combo, 1, 0, 1, 1);
-	gtk_label_set_mnemonic_widget(GTK_LABEL(sbp_label), m_search_bar_position_combo);
-	gtk_size_group_add_widget(label_size_group, sbp_label);
-	gtk_size_group_add_widget(control_size_group, m_search_bar_position_combo);
-
-	connect(m_search_bar_position_combo, "changed",
-		[this](GtkComboBox* combo)
-		{
-			const gchar* val = gtk_combo_box_get_active_id(combo);
-			if (!val)
-				return;
-			m_settings->search_bar_position = val;
-			m_plugin->reload_menu();
-			refresh_customized_indicator();
-		});
-
-	// =========================================================================
-	// 2. Ranking section (lifted from legacy Advanced Search tab)
-	// =========================================================================
-	{
-		GtkBox* row = GTK_BOX(gtk_box_new(GTK_ORIENTATION_HORIZONTAL, 8));
-
-		GtkWidget* lbl_fuzzy = gtk_label_new_with_mnemonic(_("_Fuzzy matching:"));
-		gtk_widget_set_halign(lbl_fuzzy, GTK_ALIGN_START);
-		gtk_widget_set_valign(lbl_fuzzy, GTK_ALIGN_CENTER);
-		gtk_box_pack_start(row, lbl_fuzzy, false, false, 0);
-
-		m_fuzzy_enabled = gtk_switch_new();
-		gtk_widget_set_halign(m_fuzzy_enabled, GTK_ALIGN_START);
-		gtk_widget_set_valign(m_fuzzy_enabled, GTK_ALIGN_CENTER);
-		gtk_switch_set_active(GTK_SWITCH(m_fuzzy_enabled),
-			static_cast<bool>(m_settings->fuzzy_enabled));
-		gtk_box_pack_start(row, m_fuzzy_enabled, false, false, 0);
-		gtk_label_set_mnemonic_widget(GTK_LABEL(lbl_fuzzy), m_fuzzy_enabled);
-
-		connect(m_fuzzy_enabled, "notify::active",
-			[this](GObject* obj, GParamSpec*)
-			{
-				const bool active = gtk_switch_get_active(GTK_SWITCH(obj));
-				m_settings->fuzzy_enabled = active;
-				gtk_widget_set_sensitive(m_fuzzy_threshold, active);
-			});
-
-		gtk_box_pack_start(row, gtk_separator_new(GTK_ORIENTATION_VERTICAL), false, false, 4);
-
-		GtkWidget* lbl_errors = gtk_label_new_with_mnemonic(_("Max _errors (0=auto):"));
-		gtk_widget_set_halign(lbl_errors, GTK_ALIGN_START);
-		gtk_widget_set_valign(lbl_errors, GTK_ALIGN_CENTER);
-		gtk_box_pack_start(row, lbl_errors, false, false, 0);
-
-		m_fuzzy_threshold = gtk_spin_button_new_with_range(0, 2, 1);
-		gtk_widget_set_halign(m_fuzzy_threshold, GTK_ALIGN_START);
-		gtk_spin_button_set_value(GTK_SPIN_BUTTON(m_fuzzy_threshold),
-			static_cast<int>(m_settings->fuzzy_threshold));
-		gtk_widget_set_sensitive(m_fuzzy_threshold,
-			static_cast<bool>(m_settings->fuzzy_enabled));
-		gtk_box_pack_start(row, m_fuzzy_threshold, false, false, 0);
-		gtk_label_set_mnemonic_widget(GTK_LABEL(lbl_errors), m_fuzzy_threshold);
-
-		connect(m_fuzzy_threshold, "value-changed",
-			[this](GtkSpinButton* btn)
-			{
-				m_settings->fuzzy_threshold = gtk_spin_button_get_value_as_int(btn);
-			});
-
-		gtk_box_pack_start(page,
-			make_info_frame(_("Fuzzy Search"), GTK_WIDGET(row),
-				_("Finds apps even when you mistype a word.\n"
-				  "Example: \"firfox\" still finds Firefox.\n"
-				  "Max errors 0 = automatic (1 for short queries, 2 for longer ones).")),
-			false, false, 0);
-	}
-
-	{
-		GtkGrid* grid = GTK_GRID(gtk_grid_new());
-		gtk_grid_set_column_spacing(grid, 12);
-		gtk_grid_set_row_spacing(grid, 6);
-
-		GtkBox* boost_row = GTK_BOX(gtk_box_new(GTK_ORIENTATION_HORIZONTAL, 8));
-
-		GtkWidget* lbl_boost = gtk_label_new_with_mnemonic(_("_Boost favorites:"));
-		gtk_widget_set_halign(lbl_boost, GTK_ALIGN_START);
-		gtk_widget_set_valign(lbl_boost, GTK_ALIGN_CENTER);
-		gtk_box_pack_start(boost_row, lbl_boost, false, false, 0);
-
-		m_favorites_boost_enabled = gtk_switch_new();
-		gtk_widget_set_halign(m_favorites_boost_enabled, GTK_ALIGN_START);
-		gtk_widget_set_valign(m_favorites_boost_enabled, GTK_ALIGN_CENTER);
-		gtk_switch_set_active(GTK_SWITCH(m_favorites_boost_enabled),
-			static_cast<bool>(m_settings->favorites_boost_enabled));
-		gtk_box_pack_start(boost_row, m_favorites_boost_enabled, false, false, 0);
-		gtk_label_set_mnemonic_widget(GTK_LABEL(lbl_boost), m_favorites_boost_enabled);
-
-		connect(m_favorites_boost_enabled, "notify::active",
-			[this](GObject* obj, GParamSpec*)
-			{
-				const bool active = gtk_switch_get_active(GTK_SWITCH(obj));
-				m_settings->favorites_boost_enabled = active;
-				gtk_widget_set_sensitive(m_favorites_boost_level, active);
-			});
-
-		gtk_box_pack_start(boost_row, gtk_separator_new(GTK_ORIENTATION_VERTICAL), false, false, 4);
-
-		GtkWidget* lbl_level = gtk_label_new_with_mnemonic(_("Boost _level:"));
-		gtk_widget_set_halign(lbl_level, GTK_ALIGN_START);
-		gtk_widget_set_valign(lbl_level, GTK_ALIGN_CENTER);
-		gtk_box_pack_start(boost_row, lbl_level, false, false, 0);
-
-		m_favorites_boost_level = gtk_combo_box_text_new();
-		gtk_widget_set_halign(m_favorites_boost_level, GTK_ALIGN_START);
-		gtk_combo_box_text_append_text(GTK_COMBO_BOX_TEXT(m_favorites_boost_level), _("Low"));
-		gtk_combo_box_text_append_text(GTK_COMBO_BOX_TEXT(m_favorites_boost_level), _("Medium"));
-		gtk_combo_box_text_append_text(GTK_COMBO_BOX_TEXT(m_favorites_boost_level), _("High"));
-		gtk_combo_box_set_active(GTK_COMBO_BOX(m_favorites_boost_level),
-			static_cast<int>(m_settings->favorites_boost_level) - 1);
-		gtk_widget_set_sensitive(m_favorites_boost_level,
-			static_cast<bool>(m_settings->favorites_boost_enabled));
-		gtk_box_pack_start(boost_row, m_favorites_boost_level, false, false, 0);
-		gtk_label_set_mnemonic_widget(GTK_LABEL(lbl_level), m_favorites_boost_level);
-
-		connect(m_favorites_boost_level, "changed",
-			[this](GtkComboBox* combo)
-			{
-				m_settings->favorites_boost_level = gtk_combo_box_get_active(combo) + 1;
-			});
-
-		gtk_grid_attach(grid, GTK_WIDGET(boost_row), 0, 0, 2, 1);
-
-		GtkWidget* lbl_recency = gtk_label_new_with_mnemonic(_("Recency _weight (%):"));
-		gtk_widget_set_halign(lbl_recency, GTK_ALIGN_START);
-		gtk_widget_set_valign(lbl_recency, GTK_ALIGN_CENTER);
-		gtk_grid_attach(grid, lbl_recency, 0, 1, 1, 1);
-
-		m_frecency_alpha = gtk_scale_new_with_range(GTK_ORIENTATION_HORIZONTAL, 0.0, 100.0, 1.0);
-		gtk_widget_set_hexpand(m_frecency_alpha, true);
-		gtk_scale_set_value_pos(GTK_SCALE(m_frecency_alpha), GTK_POS_RIGHT);
-		gtk_range_set_value(GTK_RANGE(m_frecency_alpha),
-			static_cast<int>(m_settings->frecency_alpha));
-		gtk_grid_attach(grid, m_frecency_alpha, 1, 1, 1, 1);
-		gtk_label_set_mnemonic_widget(GTK_LABEL(lbl_recency), m_frecency_alpha);
-
-		connect(m_frecency_alpha, "value-changed",
-			[this](GtkRange* range)
-			{
-				m_settings->frecency_alpha = static_cast<int>(gtk_range_get_value(range));
-			});
-
-		gtk_box_pack_start(page,
-			make_info_frame(_("Usage Boost"), GTK_WIDGET(grid),
-				_("Promotes apps you use frequently or marked as favorites.\n"
-				  "Recency weight controls the balance between how recently vs.\n"
-				  "how often you launched an app.")),
-			false, false, 0);
-	}
-
-	// =========================================================================
-	// 3. Aliases section
-	// =========================================================================
-	{
-		enum { ALIAS_COL_NAME, ALIAS_COL_TERMS, ALIAS_COL_ID, ALIAS_N_COLS };
-
-		m_aliases_model = gtk_list_store_new(ALIAS_N_COLS,
-			G_TYPE_STRING, G_TYPE_STRING, G_TYPE_STRING);
-
-		if (m_plugin->get_window())
-		{
-			const auto launchers = m_plugin->get_window()->get_applications()->find_all();
-			for (const auto* launcher : launchers)
-			{
-				const char* id = launcher->get_desktop_id();
-				const auto& terms = m_settings->get_aliases(id);
-				if (terms.empty())
-					continue;
-				std::string joined;
-				for (size_t i = 0; i < terms.size(); ++i)
-				{
-					if (i) joined += ", ";
-					joined += terms[i];
-				}
-				gtk_list_store_insert_with_values(m_aliases_model, nullptr, G_MAXINT,
-					ALIAS_COL_NAME, launcher->get_display_name(),
-					ALIAS_COL_TERMS, joined.c_str(),
-					ALIAS_COL_ID, id,
-					-1);
-			}
-		}
-
-		m_aliases_view = GTK_TREE_VIEW(gtk_tree_view_new_with_model(GTK_TREE_MODEL(m_aliases_model)));
-
-		GtkCellRenderer* renderer = gtk_cell_renderer_text_new();
-		GtkTreeViewColumn* col = gtk_tree_view_column_new_with_attributes(
-			_("Application"), renderer, "text", ALIAS_COL_NAME, nullptr);
-		gtk_tree_view_column_set_expand(col, true);
-		gtk_tree_view_append_column(m_aliases_view, col);
-
-		renderer = gtk_cell_renderer_text_new();
-		g_object_set(renderer, "editable", TRUE, nullptr);
-		col = gtk_tree_view_column_new_with_attributes(
-			_("Aliases (comma-separated)"), renderer, "text", ALIAS_COL_TERMS, nullptr);
-		gtk_tree_view_column_set_expand(col, true);
-		gtk_tree_view_append_column(m_aliases_view, col);
-
-		connect(renderer, "edited",
-			[this](GtkCellRendererText*, const gchar* path_str, const gchar* new_text)
-			{
-				enum { ALIAS_COL_NAME, ALIAS_COL_TERMS, ALIAS_COL_ID, ALIAS_N_COLS };
-				GtkTreeIter iter;
-				if (!gtk_tree_model_get_iter_from_string(GTK_TREE_MODEL(m_aliases_model),
-						&iter, path_str))
-					return;
-				gtk_list_store_set(m_aliases_model, &iter, ALIAS_COL_TERMS, new_text, -1);
-				gchar* id_val = nullptr;
-				gtk_tree_model_get(GTK_TREE_MODEL(m_aliases_model), &iter,
-					ALIAS_COL_ID, &id_val, -1);
-				if (id_val)
-				{
-					std::vector<std::string> terms;
-					gchar** parts = g_strsplit(new_text, ",", -1);
-					for (int i = 0; parts[i]; ++i)
-					{
-						gchar* stripped = g_strstrip(parts[i]);
-						if (*stripped)
-							terms.emplace_back(stripped);
-					}
-					g_strfreev(parts);
-					m_settings->set_aliases(std::string(id_val), terms);
-					g_free(id_val);
-				}
-			});
-
-		GtkWidget* scrolled = gtk_scrolled_window_new(nullptr, nullptr);
-		gtk_scrolled_window_set_policy(GTK_SCROLLED_WINDOW(scrolled),
-			GTK_POLICY_AUTOMATIC, GTK_POLICY_AUTOMATIC);
-		gtk_scrolled_window_set_shadow_type(GTK_SCROLLED_WINDOW(scrolled), GTK_SHADOW_IN);
-		gtk_widget_set_size_request(scrolled, -1, 120);
-		gtk_container_add(GTK_CONTAINER(scrolled), GTK_WIDGET(m_aliases_view));
-
-		m_alias_add = gtk_button_new_with_mnemonic(_("_Add"));
-		m_alias_remove = gtk_button_new_with_mnemonic(_("_Remove"));
-
-		connect(m_alias_add, "clicked",
-			[this](GtkButton*)
-			{
-				if (!m_plugin->get_window())
-					return;
-				enum { ALIAS_COL_NAME, ALIAS_COL_TERMS, ALIAS_COL_ID, ALIAS_N_COLS };
-				GtkWidget* dialog = gtk_dialog_new_with_buttons(
-					_("Choose Application"),
-					GTK_WINDOW(m_window),
-					GTK_DIALOG_MODAL,
-					_("_Cancel"), GTK_RESPONSE_CANCEL,
-					_("_Add"),    GTK_RESPONSE_ACCEPT,
-					nullptr);
-				GtkListStore* app_store = gtk_list_store_new(2, G_TYPE_STRING, G_TYPE_STRING);
-				const auto launchers = m_plugin->get_window()->get_applications()->find_all();
-				for (const auto* launcher : launchers)
-				{
-					gtk_list_store_insert_with_values(app_store, nullptr, G_MAXINT,
-						0, launcher->get_display_name(),
-						1, launcher->get_desktop_id(),
-						-1);
-				}
-				GtkWidget* app_view = gtk_tree_view_new_with_model(GTK_TREE_MODEL(app_store));
-				g_object_unref(app_store);
-				GtkCellRenderer* rend = gtk_cell_renderer_text_new();
-				gtk_tree_view_append_column(GTK_TREE_VIEW(app_view),
-					gtk_tree_view_column_new_with_attributes(
-						_("Application"), rend, "text", 0, nullptr));
-				GtkWidget* sw = gtk_scrolled_window_new(nullptr, nullptr);
-				gtk_scrolled_window_set_policy(GTK_SCROLLED_WINDOW(sw),
-					GTK_POLICY_AUTOMATIC, GTK_POLICY_AUTOMATIC);
-				gtk_widget_set_size_request(sw, 300, 240);
-				gtk_container_add(GTK_CONTAINER(sw), app_view);
-				gtk_box_pack_start(
-					GTK_BOX(gtk_dialog_get_content_area(GTK_DIALOG(dialog))),
-					sw, true, true, 6);
-				gtk_widget_show_all(dialog);
-				if (gtk_dialog_run(GTK_DIALOG(dialog)) == GTK_RESPONSE_ACCEPT)
-				{
-					GtkTreeSelection* sel = gtk_tree_view_get_selection(
-						GTK_TREE_VIEW(app_view));
-					GtkTreeIter it;
-					GtkTreeModel* mdl;
-					if (gtk_tree_selection_get_selected(sel, &mdl, &it))
-					{
-						gchar* name = nullptr;
-						gchar* id   = nullptr;
-						gtk_tree_model_get(mdl, &it, 0, &name, 1, &id, -1);
-						gtk_list_store_insert_with_values(m_aliases_model,
-							nullptr, G_MAXINT,
-							ALIAS_COL_NAME, name,
-							ALIAS_COL_TERMS, "",
-							ALIAS_COL_ID, id,
-							-1);
-						g_free(name);
-						g_free(id);
-					}
-				}
-				gtk_widget_destroy(dialog);
-				gtk_widget_set_sensitive(m_alias_remove, true);
-			});
-
-		connect(m_alias_remove, "clicked",
-			[this](GtkButton*)
-			{
-				enum { ALIAS_COL_NAME, ALIAS_COL_TERMS, ALIAS_COL_ID, ALIAS_N_COLS };
-				GtkTreeSelection* sel = gtk_tree_view_get_selection(m_aliases_view);
-				GtkTreeIter iter;
-				GtkTreeModel* mdl;
-				if (!gtk_tree_selection_get_selected(sel, &mdl, &iter))
-					return;
-				gchar* id_val = nullptr;
-				gtk_tree_model_get(mdl, &iter, ALIAS_COL_ID, &id_val, -1);
-				if (id_val)
-				{
-					m_settings->set_aliases(std::string(id_val), {});
-					g_free(id_val);
-				}
-				gtk_list_store_remove(m_aliases_model, &iter);
-				const bool has_rows = gtk_tree_model_iter_n_children(
-					GTK_TREE_MODEL(m_aliases_model), nullptr) > 0;
-				gtk_widget_set_sensitive(m_alias_remove, has_rows);
-			});
-
-		const bool has_rows = gtk_tree_model_iter_n_children(
-			GTK_TREE_MODEL(m_aliases_model), nullptr) > 0;
-		gtk_widget_set_sensitive(m_alias_remove, has_rows);
-
-		GtkWidget* btn_box = gtk_button_box_new(GTK_ORIENTATION_HORIZONTAL);
-		gtk_button_box_set_layout(GTK_BUTTON_BOX(btn_box), GTK_BUTTONBOX_START);
-		gtk_box_set_spacing(GTK_BOX(btn_box), 6);
-		gtk_box_pack_start(GTK_BOX(btn_box), m_alias_add, false, false, 0);
-		gtk_box_pack_start(GTK_BOX(btn_box), m_alias_remove, false, false, 0);
-
-		GtkBox* alias_vbox = GTK_BOX(gtk_box_new(GTK_ORIENTATION_VERTICAL, 6));
-		gtk_box_pack_start(alias_vbox, scrolled, true, true, 0);
-		gtk_box_pack_start(alias_vbox, btn_box, false, false, 0);
-
-		gtk_box_pack_start(page,
-			make_aligned_frame(_("Aliases"), GTK_WIDGET(alias_vbox)),
-			false, false, 0);
-	}
-
-	// =========================================================================
-	// 4. Search actions section — list with add/remove/edit (modal); the modal
-	// replaces the legacy inline detail panel per data-model E-5 / FR-031.
-	// =========================================================================
-	{
-		GtkGrid* actions_grid = GTK_GRID(gtk_grid_new());
-		gtk_grid_set_column_spacing(actions_grid, 6);
-		gtk_grid_set_row_spacing(actions_grid, 6);
-
-		m_actions_model = gtk_list_store_new(N_COLUMNS,
-			G_TYPE_STRING, G_TYPE_STRING, G_TYPE_POINTER);
-		for (auto action : m_settings->search_actions)
-		{
-			gtk_list_store_insert_with_values(m_actions_model,
-				nullptr, G_MAXINT,
-				COLUMN_NAME, action->get_name(),
-				COLUMN_PATTERN, action->get_pattern(),
-				COLUMN_ACTION, action,
-				-1);
-		}
-
-		m_actions_view = GTK_TREE_VIEW(gtk_tree_view_new_with_model(GTK_TREE_MODEL(m_actions_model)));
-
-		GtkCellRenderer* renderer = gtk_cell_renderer_text_new();
-		GtkTreeViewColumn* column = gtk_tree_view_column_new_with_attributes(_("Name"),
-			renderer, "text", COLUMN_NAME, nullptr);
-		gtk_tree_view_column_set_expand(column, true);
-		gtk_tree_view_append_column(m_actions_view, column);
-
-		renderer = gtk_cell_renderer_text_new();
-		column = gtk_tree_view_column_new_with_attributes(_("Pattern"),
-			renderer, "text", COLUMN_PATTERN, nullptr);
-		gtk_tree_view_column_set_expand(column, true);
-		gtk_tree_view_append_column(m_actions_view, column);
-
-		GtkTreeSelection* selection = gtk_tree_view_get_selection(m_actions_view);
-		gtk_tree_selection_set_mode(selection, GTK_SELECTION_BROWSE);
-
-		GtkWidget* sw = gtk_scrolled_window_new(nullptr, nullptr);
-		gtk_scrolled_window_set_policy(GTK_SCROLLED_WINDOW(sw),
-			GTK_POLICY_NEVER, GTK_POLICY_AUTOMATIC);
-		gtk_scrolled_window_set_shadow_type(GTK_SCROLLED_WINDOW(sw), GTK_SHADOW_ETCHED_IN);
-		gtk_container_add(GTK_CONTAINER(sw), GTK_WIDGET(m_actions_view));
-		gtk_widget_set_hexpand(sw, true);
-		gtk_widget_set_vexpand(sw, true);
-		gtk_widget_set_size_request(sw, -1, 160);
-		gtk_grid_attach(actions_grid, sw, 0, 0, 1, 1);
-
-		// Add / Remove / Edit buttons
-		m_action_add = gtk_button_new();
-		gtk_widget_set_tooltip_text(m_action_add, _("Add action"));
-		gtk_container_add(GTK_CONTAINER(m_action_add),
-			gtk_image_new_from_icon_name("list-add", GTK_ICON_SIZE_BUTTON));
-		connect(m_action_add, "clicked",
-			[this](GtkButton*) { add_action(); });
-
-		m_action_remove = gtk_button_new();
-		gtk_widget_set_tooltip_text(m_action_remove, _("Remove selected action"));
-		gtk_container_add(GTK_CONTAINER(m_action_remove),
-			gtk_image_new_from_icon_name("list-remove", GTK_ICON_SIZE_BUTTON));
-		connect(m_action_remove, "clicked",
-			[this](GtkButton*) { remove_action(); });
-
-		// Edit button — the "hamburger" of T024; opens the per-action modal
-		// (T025). open-menu-symbolic matches the GTK convention for hamburger.
-		GtkWidget* edit_btn = gtk_button_new();
-		gtk_widget_set_tooltip_text(edit_btn, _("Edit selected action…"));
-		gtk_container_add(GTK_CONTAINER(edit_btn),
-			gtk_image_new_from_icon_name("open-menu-symbolic", GTK_ICON_SIZE_BUTTON));
-		connect(edit_btn, "clicked",
-			[this](GtkButton*)
-			{
-				SearchAction* action = get_selected_action();
-				if (!action)
-					return;
-				edit_search_action_modal(action);
-			});
-
-		// Double-clicking a row should also open the modal (UX shortcut).
-		connect(m_actions_view, "row-activated",
-			[this](GtkTreeView*, GtkTreePath*, GtkTreeViewColumn*)
-			{
-				SearchAction* action = get_selected_action();
-				if (action)
-					edit_search_action_modal(action);
-			});
-
-		GtkBox* actions_box = GTK_BOX(gtk_box_new(GTK_ORIENTATION_VERTICAL, 6));
-		gtk_widget_set_halign(GTK_WIDGET(actions_box), GTK_ALIGN_START);
-		gtk_box_pack_start(actions_box, m_action_add, false, false, 0);
-		gtk_box_pack_start(actions_box, m_action_remove, false, false, 0);
-		gtk_box_pack_start(actions_box, edit_btn, false, false, 0);
-		gtk_grid_attach(actions_grid, GTK_WIDGET(actions_box), 1, 0, 1, 1);
-
-		// The legacy inline-detail panel is replaced by the modal; keep the
-		// detail entry widgets nullptr so callers that null-guard them stay
-		// safe and the destructor's unused-widget paths remain valid.
-		m_action_name = nullptr;
-		m_action_pattern = nullptr;
-		m_action_command = nullptr;
-		m_action_regex = nullptr;
-
-		const bool has_rows = !m_settings->search_actions.empty();
-		gtk_widget_set_sensitive(m_action_remove, has_rows);
-		gtk_widget_set_sensitive(edit_btn, has_rows);
-		if (has_rows)
-		{
-			GtkTreePath* path = gtk_tree_path_new_first();
-			gtk_tree_view_set_cursor(m_actions_view, path, nullptr, false);
-			gtk_tree_path_free(path);
-		}
-
-		// Re-evaluate Edit-button sensitivity whenever the selection toggles.
-		connect(selection, "changed",
-			[edit_btn](GtkTreeSelection* sel)
-			{
-				GtkTreeIter iter;
-				GtkTreeModel* mdl;
-				const bool sel_ok = gtk_tree_selection_get_selected(sel, &mdl, &iter);
-				gtk_widget_set_sensitive(edit_btn, sel_ok);
-			});
-
-		gtk_box_pack_start(page,
-			make_aligned_frame(_("Search actions"), GTK_WIDGET(actions_grid)),
-			true, true, 0);
-	}
-
-	return wrap_in_scrolled(GTK_WIDGET(page));
-}
 
 /* init_app_grid_tab:
  *
